Implement no type (or "") as generic and return OBJECT ("object" to be the correct json map type)
