+++ flowsamples/mandlebrot/pixel_to_point/pixel_to_point.toml
@@ -20,4 +20,3 @@ type = "array/number"
 
 # Pixel and point are output together
 [[output]]
-type = "object"
