+++ flowstdlib/control/join/join.toml
@@ -5,11 +5,8 @@ type = "rust"
 
 [[input]]
 name = "data" # - the data we wish to control the flow of
-type = "object"
 
 [[input]]
 name = "control" # - a second value we wait on
-type = "object"
 
 [[output]]
-type = "object"
