+++ flowstdlib/control/select/select.toml
@@ -5,11 +5,9 @@ type = "rust"
 
 [[input]]
 name = "i1"
-type = "object"
 
 [[input]]
 name = "i2"
-type = "object"
 
 [[input]]
 name = "control" # - a boolean value to selection of inputs passed to outputs
@@ -17,8 +15,6 @@ type = "boolean"
 
 [[output]]
 name = "select_i1" # if `control` is true `i1` is selected for this output else `i2` is selected
-type = "object"
 
 [[output]]
 name = "select_i2" # if `control` is true `i2` is selected for this output else `i1` is selected
-type = "object"
