+++ flowstdlib/control/tap/tap.toml
@@ -5,11 +5,9 @@ type = "rust"
 
 [[input]]
 name = "data" # - the data flow we wish to control the flow if
-type = "object"
 
 [[input]]
 name = "control" # - a boolean value to determine in `data` is passed on or not
 type = "boolean"
 
-[[output]]
-type = "object" # `data` is output if `control` is true, nothing if `control` is false
+[[output]] # `data` is output if `control` is true, nothing if `control` is false
