+++ flowstdlib/data/count/count.toml
@@ -5,16 +5,12 @@ type = "rust"
 
 [[input]]
 name = "data" # the value to count and pass on
-type = "object"
 
 [[input]]
 name = "count" # - the count before this data item
-type = "object"
 
 [[output]]
 name = "data" # - the value passed on
-type = "object"
 
 [[output]]
 name = "count" # - the incremented count of values passed thru
-type = "object"
