+++ flowstdlib/data/remove/remove.toml
@@ -8,7 +8,7 @@ name = "value" # - The value to remove from the array
 
 [[input]]
 name = "array" # - An array, to remove `value` from
-type = "array/object"
+type = "array/"
 
 [[output]]
-type = "array/object" # - The resulting array
+type = "array/" # - The resulting array
