+++ flowstdlib/data/zip/zip.toml
@@ -5,12 +5,12 @@ type = "rust"
 
 [[input]]
 name = "left" # - the 'left' array
-type = "array/object"
+type = "array/"
 
 [[input]]
 name = "right" # - the 'right' array
-type = "array/object"
+type = "array/"
 
 [[output]]
 name = "tuples" # - tuples - the array of tuples of (left, right)
-type = "array/object"
+type = "array/"
